// dreamavenue-main/api/src/modules/analysis/models/thirdparty.property.classes.dto.ts
export interface ThirdPartyProperty {
  id: number;
  parcelID: any;
  xCoordinate: string;
  yCoordinate: string;
  propertyFullStreetAddress: string;
  propertyCityName: string;
  propertyState: string;
  propertyZipCode: string;
  propertyUnitType: string;
  propertyUnitNumber: string;
  standardShortAddress: string;
  standardLongAddress: any;
  cad_CADCode: string;
  propertyType: string;
  assessmentRollActiveId: number;
  propertyId: number;
  fips: number;
  assessorsParcelNumber: string;
  lotSizeSquareFeet: string;
  lotSizeorArea: string;
  lotSizeOrAreaUnit: string;
  assessedLandValue: number;
  assessedImprovementValue: number;
  totalAssessedValue: number;
  totalMarketValue: number;
  marketValueLand: number;
  legalBriefDescription: string;
  buildingArea: number;
  mainBuildingAreaIndicator: string;
  yearBuilt: number;
  category: string;
  subcategory: string;
  assesseeOrOwnerName: string;
  latestSaleRecordingDate: string;
  categoryCount: any;
  categoryID: number;
  location: Location;
  imageURL: string;
  staticMapURL: any;
  totalAssessedValuePerSQFT: number;
  totalMarketValuePerSQFT: number;
  totalMarketValuePerSqFtRange: any;
  taxAccountNumber: any;
  detailPageURL: string;
  salePriceFromTransaction: number;
  saleDateFromTransaction: string;
  adjustedPSF: number;
  transactionSalesYear: any;
  transactionSalePrice: number;
  transactionSaleDate: string;
  transactionGBA: number;
  transactionLotSize: number;
  transactionSalePricePerSqFt: number;
  priorMortgageAmount: number;
  priorMortgageDate: string;
  priorMortgagePricePerSqFt: number;
  priorMortgageYear: number;
  lastUpdatedOn: string;
  debugvalue: any;
  nra: number;
  assessmentYear: number;
  numberofUnits: number;
  distance: number;
  transactionSaleDateRange: any;
  ownerAddress: any;
  currentOwnerName: any;
  saletype: any;
  displaySource: any;
  priorSaleDate: string;
  sellerName: any;
  buyerName: any;
  yearRenovated: number;
  censusTract: any;
  frontage: any;
  buildingClass: string;
  multipleproperty: boolean;
  actualOccupancy: number;
  signatureDate: string;
  potentialGrossIncome: number;
  vacancy: number;
  effectiveGrossIncome: number;
  expenses: number;
  nOI: number;
  buyerCompany: any;
  buyerAddress: any;
  buyerPhoneNumber: any;
  buyerFax: any;
  buyerEmail: any;
  sellerCompany: any;
  sellerAddress: any;
  sellerPhoneNumber: any;
  sellerFax: any;
  sellerEmail: any;
  filmcode: any;
  subDivison: string;
  transactionSalePriceRangeFrom: number;
  transactionSalePriceRangeTo: number;
  transactionSalePricePSFRangeFrom: number;
  transactionSalePricePSFRangeTo: number;
  transactionbuyerEntityName: any;
  transactionsellerEntityName: any;
  transactionbuyerName: any;
  es_Flag: boolean;
  propertyGrade: string;
  noofBuildings: number;
  noofStories: string;
  milesFromSubject: number;
  metaDescription: string;
  metaKeyWords: string;
  metaURL: string;
  salesHistory: any;
  mortgageHistory: any;
  assessmentHistory: any;
  rentRoll: any;
}

export interface Location {
  latitude: number;
  longitude: number;
}